add Run tool/menu，change something for wine

+++ BatchStudio/managed/Script.dsl
@@ -51,19 +51,66 @@ script(executecommand)args($cmdType, $cmdArgs)
             return(-1);
         };
     }
+    elseif ($cmdType=="buildengine") {
+        if ($cmdArgs=="dbgscp") {
+            fileecho(false);
+            command{
+                win{:e:/UGit/unity_engine/buildall.bat:};
+                nowait(true);
+                useshellexecute(true);
+            };
+            fileecho(true);
+        };
+    }
     elseif ($cmdType=="copyengine") {
         clrscpcon();
         if ($cmdArgs=="dev") {
             copydir("E:/UGit/unity_engine/build/AndroidPlayer/Variations/il2cpp/Development/Libs","E:/AndroidPlayerDebug/dev_memdbg/Libs");
             copydir("E:/UGit/unity_engine/build/AndroidPlayer/Variations/il2cpp/Development/Symbols","E:/AndroidPlayerDebug/dev_memdbg/Symbols");
         }
-        elseif($cmdArgs=="rel") {
+        elseif ($cmdArgs=="rel") {
             copydir("E:/UGit/unity_engine/build/AndroidPlayer/Variations/il2cpp/Release/Libs","E:/AndroidPlayerDebug/rel_memdbg/Libs");
             copydir("E:/UGit/unity_engine/build/AndroidPlayer/Variations/il2cpp/Release/Symbols","E:/AndroidPlayerDebug/rel_memdbg/Symbols");
         };
         logscpcon();
+    }
+    elseif ($cmdType=="unzipsym") {
+        if ($cmdArgs=="testuc") {
+            $args = format("x {0}/{1} -o{0} -y", "e:/UGit/TestUC", "testuc-0.1-v1-IL2CPP.symbols.zip");
+            $r = nativeapi.RunCommand("c:/Program Files/7-Zip/7z.exe", $args);
+            if (!$r) {
+                outputlog("unzip sym failed:{0}", nativeapi.GetResultCode());
+                return(-1);
+            };
+        };
+    }
+    elseif ($cmdType=="open") {
+        if ($cmdArgs=="testuc") {
+            process("e:/UGit/unity_engine/build/WindowsEditor/x64/Release/Unity.exe", "-projectPath e:/UGit/TestUC"){
+                nowait(true);
+            };
+        }
+        elseif ($cmdArgs=="androidstudio") {
+            process("C:/Program Files/Android/Android Studio/bin/studio64.exe", ""){
+                nowait(true);
+            };
+        }
+        elseif ($cmdArgs=="quickmirror") {
+            process("c:/Users/dreamanlan/AppData/Local/QuickMirror/QuickMirror.exe", ""){
+                nowait(true);
+            };
+        };
     };
     outputlog("[dsl] execute command success");
+    return(1);
+};
+
+script(runprog)
+{
+    $path = getdirectoryname(selInTree);
+    $filename = getfilename(selInTree);
+
+    outputlog("[dsl] running program success");
     return(1);
 };
 
@@ -91,33 +138,62 @@ script(build)
 
 script(install)
 {
+    $filename = getfilename(selInTree);
     $ext = getextension(selInTree);
-    if ($ext==".apk") {
-        $adbExe = nativeapi.GetAdbExe();
-        if (isnullorempty($adbExe)) {
-            return(-1);
-        };
-        $args = format("install {0}", selInTree);
-        $r = nativeapi.RunCommand($adbExe, $args);
-        if (!$r) {
-            outputlog("install failed:{0}", nativeapi.GetResultCode());
-            return(-1);
+    if (direxist(selInTree)) {
+        if ($filename=="AssetsBundle") {
+            $adbExe = nativeapi.GetAdbExe();
+            if (isnullorempty($adbExe)) {
+                return(-1);
+            };
+            $args = format("push {0} /sdcard/Android/data/com.DefaultCompany.TestUC05/files/", selInTree);
+            $r = nativeapi.RunCommand($adbExe, $args);
+            if (!$r) {
+                outputlog("install failed:{0}", nativeapi.GetResultCode());
+                return(-1);
+            };
         };
     }
-    elseif ($ext==".so"){
-        $adbExe = nativeapi.GetAdbExe();
-        if (isnullorempty($adbExe)) {
-            return(-1);
-        };
-        $args = format("shell mkdir /sdcard/Android/data/com.tencent.uc/files/ExtraFiles");
-        $r = nativeapi.RunCommand($adbExe, $args);
-        $args = format("shell mkdir /sdcard/Android/data/com.tencent.uc/files/ExtraFiles/arm64-v8a");
-        $r = nativeapi.RunCommand($adbExe, $args);
-        $args = format("push {0} /sdcard/Android/data/com.tencent.uc/files/ExtraFiles/arm64-v8a/", selInTree);
-        $r = nativeapi.RunCommand($adbExe, $args);
-        if (!$r) {
-            outputlog("install failed:{0}", nativeapi.GetResultCode());
-            return(-1);
+    else{
+        if ($ext==".apk") {
+            $adbExe = nativeapi.GetAdbExe();
+            if (isnullorempty($adbExe)) {
+                return(-1);
+            };
+            $args = format("install {0}", selInTree);
+            $r = nativeapi.RunCommand($adbExe, $args);
+            if (!$r) {
+                outputlog("install failed:{0}", nativeapi.GetResultCode());
+                return(-1);
+            };
+        }
+        elseif ($ext==".so") {
+            $adbExe = nativeapi.GetAdbExe();
+            if (isnullorempty($adbExe)) {
+                return(-1);
+            };
+            $args = format("shell mkdir /sdcard/Android/data/com.tencent.uc/files/ExtraFiles");
+            $r = nativeapi.RunCommand($adbExe, $args);
+            $args = format("shell mkdir /sdcard/Android/data/com.tencent.uc/files/ExtraFiles/arm64-v8a");
+            $r = nativeapi.RunCommand($adbExe, $args);
+            $args = format("push {0} /sdcard/Android/data/com.tencent.uc/files/ExtraFiles/arm64-v8a/", selInTree);
+            $r = nativeapi.RunCommand($adbExe, $args);
+            if (!$r) {
+                outputlog("install failed:{0}", nativeapi.GetResultCode());
+                return(-1);
+            };
+        }
+        else {
+            $adbExe = nativeapi.GetAdbExe();
+            if (isnullorempty($adbExe)) {
+                return(-1);
+            };
+            $args = format("push {0} /data/local/tmp/", selInTree);
+            $r = nativeapi.RunCommand($adbExe, $args);
+            if (!$r) {
+                outputlog("install failed:{0}", nativeapi.GetResultCode());
+                return(-1);
+            };
         };
     };
     outputlog("[dsl] install success");
